added cost function to all beverages and decorators

+++ src/Beverage.java
@@ -1,3 +1,3 @@
 public abstract class Beverage {
-    public abstract int cost();
+    public abstract double cost();
 }
